Shuffle around the php-encryption library setup
