--- libraries/defuse/php-encryption/.travis.yml
@@ -1,15 +1,0 @@
-language: php
-php:
-    - "5.5"
-    - "5.4"
-    - "5.3"
-    - "5.2"
-# Versions below here are not installed on travis-ci
-#    - "5.1"
-#    - "5.0"
-#    - "4.4"
-#    - "4.3"
-#    - "4.2"
-#    - "4.1"
-#    - "4.0"
-script: ./test.sh
